[fixed] modal styling
