+++ sd20s-web/src/app/globals.css
@@ -58,19 +58,37 @@ body {
 }
 
 
-.modal-overlay {
-  position: fixed;
+.header {
+  position: sticky;
   top: 0;
   left: 0;
   right: 0;
-  bottom: 0;
-  display: flex;
+  width: 100%;
+  height: 115px;
+  direction: flex;
+  flex-direction: row;
+  border-radius: 20px;
+  background-color: black;
+  z-index: 50;
   justify-content: center;
   align-items: center;
+  padding: 0px, 20px;
+
+}
+.modal-overlay {
+  position: fixed;
+  top: 115px;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: #000000;
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+  z-index: 0;
   opacity: 0;
   visibility: hidden;
-  background: rgba(0, 0.5, 0, 0.5);
-  z-index:1000;
   transition: opacity 0.3s ease-in-out, visibility 0.3s ease-in-out;
 }
 .modal-container {
@@ -79,15 +97,16 @@ body {
   border-radius: 8px;
   padding: 20px;
   width: 298px;
-  min-height: 500px;
+  min-height: 530px;
   height:auto;/* Allow the height to adjust based on content */
-  max-height: 90vh; /* Ensure the modal doesn't exceed the viewport height */
+  max-height: 50vh; /* Ensure the modal doesn't exceed the viewport height */
   overflow-y: auto; /* Add scroll if content exceeds max-height */
   overflow-x: hidden; /* Hide the horizontal scrollbar */
   display: flex;
   flex-direction: column;
   /* justify-content: center; */
   align-items: center;
+  z-index: 10;
 }
 
 .modal-overlay.active {
@@ -95,15 +114,10 @@ body {
   visibility: visible;
 }
 
-.modal-overlay.inactive{
-  display:none;
+.modal-overlay.inactive {
+  opacity: 0;
   visibility:hidden;
  }
-
-.modal-overlay.active .modal-content {
-  transform: scale(1);
-  opacity: 1;
-}
 
 .modal-close-button {
   position: absolute;
@@ -111,3 +125,15 @@ body {
   cursor: pointer;
   color: #606060;
 }
+
+.soft-star-container {
+  position: relative;
+  margin-top: 70px;
+}
+
+@media (min-width: 768px) {
+  .soft-star-container {
+    margin-bottom: 40px;
+  }
+}
+
